pyproject.toml
[project]
name = "meilleure-copro"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "ruff (>=0.14.11,<0.15.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poertry]
packages = [{include = "meilleurecopro"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
